requirements.txt
aiofiles==23.2.1
aiosqlite==0.17.0
async-timeout==4.0.3
asyncpg==0.28.0
html5tagger==1.3.0
httptools==0.6.1
iso8601==1.1.0
itsdangerous==2.0.1
multidict==6.0.4
numpy==1.24.4
pandas==2.0.3
psycopg2==2.9.6
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.4
redis==4.6.0
sanic==23.3.0
sanic-routing==23.12.0
six==1.16.0
tortoise-orm==0.20.0
tracerite==1.1.1
typing_extensions==4.9.0
tzdata==2023.4
websockets==12.0
Sanic-Auth==0.3.0
